% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildcodon.R
\name{buildcodon}
\alias{buildcodon}
\title{buildcodon}
\usage{
buildcodon(refcds, numcode = 1)
}
\arguments{
\item{refcds}{Input RefCDS object}

\item{numcode}{NCBI genetic code number (default = 1; standard genetic code). To see the list of genetic codes supported use: ? seqinr::translate}
}
\description{
Function to build a refcodon object from a RefCDS object. This function takes a RefCDS object as input and adds to it two fields required to run the codondnds function. Recommended usage: RefCDS = buildcodon(RefCDS)
}
\details{
Martincorena I, et al. (2017) Universal patterns of selection in cancer and somatic tissues. Cell. 171(5):1029-1041.
}
\author{
Inigo Martincorena (Wellcome Sanger Institute)
}
