\name{dndscv-package}
\alias{dndscv-package}
\alias{dndscv}
\docType{package}
\title{
\packageTitle{dndscv}
}
\description{
\packageDescription{dndscv}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{dndscv}
\packageIndices{dndscv}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{dndscv}

Maintainer: \packageMaintainer{dndscv}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
