% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitlnpbin.R
\name{fitlnpbin}
\alias{fitlnpbin}
\title{fitlnpbin}
\usage{
fitlnpbin(nvec, rvec, level = 0.95, theta_option = "conservative",
  numbins = 10000)
}
\arguments{
\item{nvec}{Vector of observed counts of mutations per site.}

\item{rvec}{Vector of expected counts of mutations per site.}

\item{level}{Confidence level desired for the confidence interval of the overdispersion parameter [defaul=0.95]}

\item{theta_option}{2 options: "mle" (uses the MLE of the overdispersion parameter) or "conservative" (uses the conservative bound of the CI95). Values other than "mle" will lead to the conservative option [default="conservative"]}

\item{numbins}{Number of bins to discretise the rvec vector [default=1e4]. This enables fast execution of the LNP model in datasets of arbitrarily any size.}
}
\value{
'fitlnpbin' returns the maximum likelihood estimate and confidence intervals of the "sig" overdispersion parameter of the LNP model:
}
\description{
Function to fit a Lognormal-Poisson model to estimate overdispersion on synonymous changes for sitednds and codondnds.
}
\author{
Inigo Martincorena (Wellcome Sanger Institute)
}
