% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneciloc.R
\name{geneciloc}
\alias{geneciloc}
\title{geneciloc}
\usage{
geneciloc(dndsout, gene_list = NULL, level = 0.95)
}
\arguments{
\item{dndsout}{Output object from dndscv.}

\item{gene_list}{List of genes to restrict the analysis (by default, all genes in dndsout will be analysed)}

\item{level}{Confidence level desired [default = 0.95]}
}
\value{
ci: Dataframe with the confidence intervals for dN/dS ratios per gene under the dNdSloc model.
}
\description{
Function to calculate confidence intervals for dN/dS values per gene under the dNdSloc model using a Poisson ratio.
}
\details{
Martincorena I, et al. (2017) Universal patterns of selection in cancer and somatic tissues. Cell. 171(5):1029-1041.
}
\author{
Inigo Martincorena (Wellcome Sanger Institute)
}
