% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetdnds.R
\name{genesetdnds}
\alias{genesetdnds}
\title{genesetdnds}
\usage{
genesetdnds(dndsout, gene_list, sm = "192r_3w")
}
\arguments{
\item{dndsout}{Output object from dndscv. To generate a valid input object for this function, use outmats=T when running dndscv.}

\item{gene_list}{List of genes to restrict the analysis (gene set).}

\item{sm}{Substitution model (precomputed models are available in the data directory)}
}
\value{
'genesetdnds' returns a list of objects:

globaldnds_geneset: Global dN/dS estimates in the gene set.

globaldnds_rest: Global dN/dS estimates in all other genes.
}
\description{
Function to estimate global dN/dS values for a gene set when using whole-exome 
data. Global dN/dS values for a set of genes can also be obtained using dndscv 
(gene_list argument), but that option estimates the trinucleotide mutation rates 
exclusively from the list of genes of interest. This may be prefereable in large
datasets, but in small datasets, the genesetdnds option estimates global dN/dS
values for a set of genes while using all genes in the exome to fit the 
substitution model. Usage: genesetdnds(dndsout, gene_list).
}
\details{
Martincorena I, et al. (2017) Universal patterns of selection in cancer and somatic tissues. Cell. 171(5):1029-1041.
}
\author{
Inigo Martincorena (Wellcome Sanger Institute)
}
