% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getsitedc.R
\name{getsitedc}
\alias{getsitedc}
\title{getsitedc}
\usage{
getsitedc(refcds, gr_genes, coverage_file)
}
\arguments{
\item{refcds}{Input refcds object}

\item{gr_genes}{Input gr_genes object available from the precomputed RefCDS files or generated by buildref}

\item{coverage_file}{Path to a bigwig file with the aggregate duplex coverage per site for all samples used to call mutations in the dataset}
}
\description{
This function takes a RefCDS object and a bigwig file with duplex coverage as input and adds to the RefCDS two fields required to use site-level duplex coverage in dndscv. Usage: RefCDS = getsitedc(RefCDS)
}
\details{
Martincorena I, et al. (2017) Universal patterns of selection in cancer and somatic tissues. Cell. 171(5):1029-1041.
}
\author{
Inigo Martincorena (Wellcome Sanger Institute)
}
