% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codondnds.R
\name{codondnds}
\alias{codondnds}
\title{codondnds}
\usage{
codondnds(dndsout, refcds, min_recurr = 2, gene_list = NULL,
  site_list = NULL, theta_option = "mle", syn_drivers = "TP53:T125T",
  method = "NB", numbins = 10000)
}
\arguments{
\item{dndsout}{Output object from dndscv.}

\item{refcds}{RefCDS object annotated with codon-level information using the buildcodon function.}

\item{min_recurr}{Minimum number of mutations per site to estimate codon-wise dN/dS ratios. [default=2]}

\item{gene_list}{List of genes to restrict the p-value and q-value calculations (Restricted Hypothesis Testing). Note that q-values are only valid if the list of genes is decided a priori. [default=NULL, codondnds will be run on all genes in dndsout]}

\item{site_list}{List of hotspot sites to restrict the p-value and q-value calculations (Restricted Hypothesis Testing). Note that q-values are only valid if the list of codons is decided a priori. [default=NULL, sitednds will be run on all genes in dndsout]}

\item{theta_option}{2 options: "mle" (uses the MLE of the negative binomial size parameter) or "conservative" (uses the lower bound of the CI95). Values other than "mle" will lead to the conservative option. [default="mle"]}

\item{syn_drivers}{Vector with a list of known synonymous driver mutations to exclude from the background model [default="TP53:T125T"]. See Martincorena et al., Cell, 2017 (PMID:29056346).}

\item{method}{Overdispersion model: NB = Negative Binomial (Gamma-Poisson), LNP = Poisson-Lognormal (see Hess et al., BiorXiv, 2019). [default="NB"]}

\item{numbins}{Number of bins to discretise the rvec vector [default=1e4]. This enables fast execution of the LNP model in datasets of arbitrarily any size.}
}
\value{
'codondnds' returns a table of recurrently mutated codons and the estimates of the size parameter:

- recurcodons: Table of recurrently mutated codons with codon-wise dN/dS values and p-values

- recurcodons_ext: The same table of recurrently mutated codons, but including additional information on the contribution of different changes within a codon.

- theta: Maximum likelihood estimate and CI95% for the size parameter of the negative binomial distribution. The lower this value the higher the variation of the mutation rate across sites not captured by the trinucleotide change or by variation across genes.
}
\description{
Function to estimate codon-wise dN/dS values and p-values against neutrality. To generate a valid RefCDS input object for this function, use the buildcodon function. Note that recurrent artefacts or SNP contamination can violate the null model and dominate the list of sites under apparent selection. Be very critical of the results and if suspicious sites appear recurrently mutated consider refining the variant calling (e.g. using a better unmatched normal panel).
}
\author{
Inigo Martincorena (Wellcome Sanger Institute)
}
