% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dndscv.R
\name{dndscv}
\alias{dndscv}
\title{dNdScv}
\usage{
dndscv(mutations, gene_list = NULL, refdb = "hg19",
  substmodel = "submod_192r_3w", known_cancergenes = "hg19",
  covs = "hg19", max_muts_per_gene_per_sample = 3,
  max_coding_muts_per_sample = 3000, use_indel_sites = T, min_indels = 5,
  maxcovs = 20, constrain_wnon_wspl = T)
}
\arguments{
\item{mutations}{Table of mutations (5 columns: sampleID, chr, pos, ref, alt). Only list independent events as mutations.}

\item{gene_list}{List of genes to restrict the analysis (use for targeted sequencing studies)}

\item{refdb}{Reference database (path to .rda file)}

\item{substmodel}{Substitution model (precomputed models are available in the data directory)}

\item{known_cancergenes}{List of a-priori known cancer genes (to be excluded from the indel background model)}

\item{max_muts_per_gene_per_sample}{If n<Inf, arbitrarily the first n mutations by chr position will be kept}

\item{max_coding_muts_per_sample}{Hypermutator samples often reduce power to detect selection}

\item{use_indel_sites}{Use unique indel sites instead of the total number of indels (it tends to be more robust)}

\item{min_indels}{Minimum number of indels required to run the indel recurrence module}

\item{maxcovs}{Maximum number of covariates that will be considered (additional columns in the matrix of covariates will be excluded)}

\item{constrain_wnon_wspl}{This constrains wnon==wspl (this typically leads to higher power to detect selection)}
}
\value{
'dndscv' returns a list of objects:

- globaldnds: Global dN/dS estimates across all genes.

- sel_cv: Gene-wise selection results using dNdScv.

- sel_loc: Gene-wise selection results using dNdSloc.

- annotmuts: Annotated coding mutations.

- genemuts: Observed and expected numbers of mutations per gene.

- mle_submodel: MLEs of the substitution model.

- exclsamples: Samples excluded from the analysis.

- exclmuts: Coding mutations excluded from the analysis.
}
\description{
Analyses of selection using the dNdScv and dNdSloc models. Default parameters typically increase the performance of the method on cancer genomic studies. Reference files are currently only available for the GRCh37/hg19 version of the human genome.
}
\details{
Martincorena I, et al. (2017) Universal patterns of selection in cancer and somatic tissues. Under revision. Preprint available in BioRxiv: https://doi.org/10.1101/132324
}
\author{
Inigo Martincorena (Wellcome Trust Sanger Institute)
}
