% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitednds.R
\name{sitednds}
\alias{sitednds}
\title{sitednds}
\usage{
sitednds(dndsout, min_recurr = 2, gene_list = NULL, theta_option = "mle",
  syn_drivers = "TP53:T125T")
}
\arguments{
\item{dndsout}{Output object from dndscv. To generate a valid input object for this function, use outmats=T when running dndscv.}

\item{min_recurr}{Minimum number of mutations per site to estimate site-wise dN/dS ratios. [default=2]}

\item{gene_list}{List of genes to restrict the analysis (only needed if the user wants to restrict the analysis to a subset of the genes in dndsout) [default=NULL, sitednds will be run on all genes in dndsout]}

\item{theta_option}{2 options: "mle" (uses the MLE of the negative binomial size parameter) or "conservative" (uses the lower bound of the CI95). Values other than "mle" will lead to the conservative option. [default="mle"]}

\item{syn_drivers}{Vector with a list of known synonymous driver mutations to exclude from the background model [default="TP53:T125T"]. See Martincorena et al., Cell, 2017 (PMID:29056346).}
}
\value{
'sitednds' returns a table of recurrently mutated sites and the estimates of the size parameter:

- recursites: Table of recurrently mutated sites with site-wise dN/dS values and p-values

- theta: Maximum likelihood estimate and CI95% for the size parameter of the negative binomial distribution. The lower this value the higher the variation of the mutation rate across sites not captured by the trinucleotide change or by variation across genes.
}
\description{
Function to estimate site-wise dN/dS values and p-values against neutrality. This function is in testing, please interpret the results with caution. Also note that recurrent artefacts or SNP contamination can violate the null model and dominate the list of sites under apparent selection. A considerable number of significant synonymous sites may reflect a problem with the data. Be very critical of the results and if suspicious sites appear recurrently mutated consider refining the variant calling (e.g. using a better unmatched normal panel). In the future, this function may be extended to perform inferences at a codon level instead of at a single-base level.
}
\author{
Inigo Martincorena (Wellcome Sanger Institute)
}
